requirements.txt
# Backend dependencies
fastapi>=0.95
uvicorn[standard]>=0.18
requests>=2.28
chromadb>=0.4
PyPDF2>=3.0
sentence-transformers>=2.2
python-dotenv>=1.0
python-multipart>=0.0.6
redis>=4.5.4
